Fix crash by skipping registry of weird plank/slab recipes (#167)
Fixes #166

Add workaround to skip generating plank/slab recipes using a saw or cutting machine when the default recipe has some bizarre properties that previously caused a crash. Namely, recipes with no ingredients or no matching ingredients.

When a recipe is skipped for these reasons we log a warning with the Registry Name of the skipped recipe so you can tell the appropriate people they're doing something weird (not our fault).

Also switches to using WynPrice's CurseMaven to get Forestry dependencies via CurseForge rather than the IC2 Maven since they seem to be having issues and it was causing the build to fail.